+++ changeLog.md
@@ -1,6 +1,10 @@
 
 ### caracteristicas nuevas
 
+- la bandejaEntrada pasa a ser una extension de la clase empleado
+
+- bandeja de entrada implementada (falta comprobar su funcionamiento)
+
 - la clase menuTools se encuentra en el Menu.java
 
 - se han movido algunas clases de carpeta y se han creado otras como MenuTools, que ayuda a la correcta ejecucion del programa y para organizar el codigo